Refactor location of HttpEasyConfigurator
